Remake non-functional playback in WASAPI driver 2
This properly handles the case of the output device disappearing and reappearing, or changing at runtime, or there being several output devices active at once. This logic never seem to have worked before? But it is implemented now. In case there are two audio outputs in the system and one is removed, Windows will make the remaining device the default, and this logic catches that. If the only remaining output device is removed as well, there will not be any processing until a working device is added back in to the system.

This also separates capture and rendering into separate threads because the capture logic is flawed, but it has not been touched here. So it is kept isolated to prioritize the playback case.